Introduced a prcessorManager and make sure that our processor classes remain thin
