+++ DizdaCloudBackupBundle.php
@@ -4,6 +4,7 @@ namespace Dizda\CloudBackupBundle;
 
 use Dizda\CloudBackupBundle\DependencyInjection\Compiler\ClientCompilerPass;
 use Dizda\CloudBackupBundle\DependencyInjection\Compiler\DatabaseCompilerPass;
+use Dizda\CloudBackupBundle\DependencyInjection\Compiler\ProcessorCompilerPass;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\HttpKernel\Bundle\Bundle;
 
@@ -17,5 +18,6 @@ class DizdaCloudBackupBundle extends Bundle
     {
         $container->addCompilerPass(new DatabaseCompilerPass());
         $container->addCompilerPass(new ClientCompilerPass());
+        $container->addCompilerPass(new ProcessorCompilerPass());
     }
 }
