+++ Processors/ProcessorInterface.php
@@ -10,16 +10,6 @@ namespace Dizda\CloudBackupBundle\Processors;
 interface ProcessorInterface
 {
     /**
-     * Compress to file with name like : hostname_2013-01-12_00-06-40.tar
-     */
-    public function compress();
-
-    /**
-     * Remove all dirs with files
-     */
-    public function cleanUp();
-
-    /**
      * Get compression command
      *
      * @param string $archivePath
@@ -30,21 +20,9 @@ interface ProcessorInterface
     public function getCompressionCommand($archivePath, $basePath);
 
     /**
-     * Get file extention (with leading dot)
+     * Get file extension (with leading dot)
      *
      * @return string
      */
     public function getExtension();
-
-    /**
-     * Return path of the archive
-     *
-     * @return string
-     */
-    public function getArchivePath();
-
-    /**
-     * Make a copy of all folders specified in config
-     */
-    public function copyFolders();
 }
